+++ platform/android/ndk/jni/JavaToNativeBridge.cpp
@@ -1,9 +1,25 @@
 //////////////////////////////////////////////////////////////////////////////
 //
-// This file is part of the Corona game engine.
-// For overview and more information on licensing please refer to README.md 
-// Home page: https://github.com/coronalabs/corona
+// Copyright (C) 2018 Corona Labs Inc.
 // Contact: support@coronalabs.com
+//
+// This file is part of the Corona game engine.
+//
+// Commercial License Usage
+// Licensees holding valid commercial Corona licenses may use this file in
+// accordance with the commercial license agreement between you and 
+// Corona Labs Inc. For licensing terms and conditions please contact
+// support@coronalabs.com or visit https://coronalabs.com/com-license
+//
+// GNU General Public License Usage
+// Alternatively, this file may be used under the terms of the GNU General
+// Public license version 3. The license is as published by the Free Software
+// Foundation and appearing in the file LICENSE.GPL3 included in the packaging
+// of this file. Please review the following information to ensure the GNU 
+// General Public License requirements will
+// be met: https://www.gnu.org/licenses/gpl-3.0.html
+//
+// For overview and more information on licensing please refer to README.md
 //
 //////////////////////////////////////////////////////////////////////////////
 
@@ -1122,38 +1138,38 @@ JavaToNativeBridge::AlertCallback( int which, bool cancelled )
 }
 
 void
-JavaToNativeBridge::SoundEndCallback( long id )
+JavaToNativeBridge::SoundEndCallback( int id )
 {
 	if (fNativeToJavaBridge != NULL)
 	{
-		fNativeToJavaBridge->SoundEndCallback((uintptr_t)id);
+		fNativeToJavaBridge->SoundEndCallback(id);
 	}
 }
 
 void
-JavaToNativeBridge::VideoEndCallback( long id )
+JavaToNativeBridge::VideoEndCallback( int id )
 {
 	if (fNativeToJavaBridge != NULL)
 	{
-		fNativeToJavaBridge->VideoEndCallback((uintptr_t)id);
+		fNativeToJavaBridge->VideoEndCallback(id);
 	}
 }
 
 void 
-JavaToNativeBridge::RecordCallback( long id, int status )
+JavaToNativeBridge::RecordCallback( int id, int status )
 {
 	if (fNativeToJavaBridge != NULL)
 	{
-		fNativeToJavaBridge->RecordCallback((uintptr_t)id, status);
+		fNativeToJavaBridge->RecordCallback(id, status);
 	}
 }
 
 void
-JavaToNativeBridge::SetAudioRecorderState( long id, bool isRecording )
+JavaToNativeBridge::SetAudioRecorderState( int id, bool isRecording )
 {
 	if (fPlatform && fRuntime)
 	{
-		Rtt::AndroidAudioRecorder *audioRecorderPointer = (Rtt::AndroidAudioRecorder*) (uintptr_t)id;
+		Rtt::AndroidAudioRecorder *audioRecorderPointer = (Rtt::AndroidAudioRecorder*) id;
 		if (audioRecorderPointer)
 		{
 			audioRecorderPointer->SetRunningState( isRecording );
